Refactor(InputUtils): 입력 재시도 클래스 유틸 클래스로 변경
